requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
pydantic>=2.7.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
openai>=1.40.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
docx2txt>=0.8
python-docx>=1.2.0
